Add default resource generator for log entries
To accommodate invocations like:
  `syslog.entry({textPayload: 'test'})`
Create a basic class which attempts to automatically
establish resource values based on connectivity to
the metadata service and environmental variables.

This addition does not affect the existing API and
does not modify or augment user supplied resources.
The meta data class will attempt to determine if the
resource is a cloud function, app engine application
or compute engine instance. If it can't match on any
of these configurations but can still locate a project
id then it will produce a global resource id.